backend/requirements.txt
fastapi
uvicorn
azure-functions
nest-asyncio
pandas
scikit-learn
requests
asyncio
pickle-mixin
numpy
